Override site.github.url locally:
My hope was that Github Pages would override site.github values
on deployment, making it safe to set

github:
  url: whatever

in _config.yml. However, this doesn't seem to be the case. Instead,
I created a separate config file, config.development.yml, that contains
the local override.

Them, I added foreman and a Procfile that runs jekyll serve with the
override yml.